refactor: 운동 게스트 컨트롤러 분리
