+++ src/main/java/umc/cockple/demo/domain/exercise/controller/ExerciseController.java
@@ -3,12 +3,10 @@ package umc.cockple.demo.domain.exercise.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.format.annotation.DateTimeFormat;
-import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import umc.cockple.demo.domain.exercise.dto.*;
@@ -16,7 +14,6 @@ import umc.cockple.demo.domain.exercise.enums.MyExerciseFilterType;
 import umc.cockple.demo.domain.exercise.enums.MyExerciseOrderType;
 import umc.cockple.demo.domain.exercise.enums.MyPartyExerciseOrderType;
 import umc.cockple.demo.domain.exercise.service.*;
-import umc.cockple.demo.domain.exercise.service.command.ExerciseCommandService;
 import umc.cockple.demo.domain.party.enums.ActivityTime;
 import umc.cockple.demo.domain.party.enums.ParticipationType;
 import umc.cockple.demo.global.enums.Level;
@@ -34,62 +31,7 @@ import java.util.List;
 @Tag(name = "Exercise", description = "운동 관리 API")
 public class ExerciseController {
 
-    private final ExerciseCommandService exerciseCommandService;
     private final ExerciseQueryService exerciseQueryService;
-
-    @PostMapping("/exercises/{exerciseId}/guests")
-    @Operation(summary = "게스트 초대",
-            description = "파티 멤버가 게스트를 운동에 초대합니다. 운동의 게스트 허용 정책을 확인합니다.")
-    @ApiResponse(responseCode = "201", description = "게스트 초대 성공")
-    @ApiResponse(responseCode = "400", description = "입력값 오류 또는 비즈니스 룰 위반")
-    @ApiResponse(responseCode = "404", description = "운동을 찾을 수 없음")
-    public ResponseEntity<BaseResponse<ExerciseGuestInviteDTO.Response>> inviteGuest(
-            @PathVariable Long exerciseId,
-            @Valid @RequestBody ExerciseGuestInviteDTO.Request request
-    ) {
-        Long inviterId = SecurityUtil.getCurrentMemberId();
-
-        ExerciseGuestInviteDTO.Response response = exerciseCommandService.inviteGuest(
-                exerciseId, inviterId, request);
-
-        return BaseResponse.of(CommonSuccessCode.CREATED, response);
-    }
-
-    @DeleteMapping("/exercises/{exerciseId}/guests/{guestId}")
-    @Operation(summary = "게스트 초대 취소",
-            description = "사용자가 본인이 초대한 게스트를 취소합니다.")
-    @ApiResponse(responseCode = "200", description = "게스트 초대 취소 성공")
-    @ApiResponse(responseCode = "400", description = "취소할 수 없는 상태 (이미 시작됨)")
-    @ApiResponse(responseCode = "403", description = "본인이 초대한 게스트가 아닌 경우 취소할 수 없음")
-    @ApiResponse(responseCode = "404", description = "운동 또는 참여 기록을 찾을 수 없음")
-    public ResponseEntity<BaseResponse<ExerciseCancelDTO.Response>> cancelGuestInvitation(
-            @PathVariable Long exerciseId,
-            @PathVariable Long guestId
-    ) {
-        Long memberId = SecurityUtil.getCurrentMemberId();
-
-        ExerciseCancelDTO.Response response = exerciseCommandService.cancelGuestInvitation(
-                exerciseId, guestId, memberId);
-
-        return BaseResponse.of(CommonSuccessCode.OK, response);
-    }
-
-    @GetMapping("/exercises/{exerciseId}/guests")
-    @Operation(summary = "내가 초대한 운동 게스트 조회",
-            description = "내가 초대한 운동 게스트 목록을 조회합니다.")
-    @ApiResponse(responseCode = "200", description = "내가 초대한 운동 게스트 조회 성공")
-    @ApiResponse(responseCode = "404", description = "존재하지 않는 운동")
-    public BaseResponse<ExerciseMyGuestListDTO.Response> getMyInvitedGuests(
-            @PathVariable Long exerciseId
-    ) {
-        Long memberId = SecurityUtil.getCurrentMemberId();
-
-        ExerciseMyGuestListDTO.Response response = exerciseQueryService.getMyInvitedGuests(
-                exerciseId, memberId);
-
-        return BaseResponse.success(CommonSuccessCode.OK, response);
-    }
-
 
     @GetMapping("/parties/{partyId}/exercises/calender")
     @Operation(summary = "모임 운동 캘린더 조회",
