+++ src/test/java/umc/cockple/demo/domain/exercise/integration/ExerciseGuestIntegrationTest.java
@@ -1,17 +1,22 @@
 package umc.cockple.demo.domain.exercise.integration;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.domain.exercise.domain.Guest;
 import umc.cockple.demo.domain.exercise.dto.ExerciseGuestInviteDTO;
 import umc.cockple.demo.domain.exercise.exception.ExerciseErrorCode;
 import umc.cockple.demo.domain.exercise.repository.ExerciseRepository;
 import umc.cockple.demo.domain.exercise.repository.GuestRepository;
-import umc.cockple.demo.domain.member.repository.MemberExerciseRepository;
 import umc.cockple.demo.domain.member.domain.Member;
+import umc.cockple.demo.domain.member.repository.MemberExerciseRepository;
 import umc.cockple.demo.domain.member.repository.MemberPartyRepository;
 import umc.cockple.demo.domain.member.repository.MemberRepository;
 import umc.cockple.demo.domain.party.domain.Party;
@@ -23,22 +28,21 @@ import umc.cockple.demo.global.enums.Level;
 import umc.cockple.demo.global.enums.Role;
 import umc.cockple.demo.support.IntegrationTestBase;
 import umc.cockple.demo.support.SecurityContextHelper;
-import umc.cockple.demo.support.fixture.MemberFixture;
-import umc.cockple.demo.support.fixture.PartyFixture;
-
-import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.support.fixture.ExerciseFixture;
 import umc.cockple.demo.support.fixture.GuestFixture;
+import umc.cockple.demo.support.fixture.MemberFixture;
+import umc.cockple.demo.support.fixture.PartyFixture;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-class ExerciseCommandIntegrationTest extends IntegrationTestBase {
+class ExerciseGuestIntegrationTest extends IntegrationTestBase {
 
     @Autowired MockMvc mockMvc;
     @Autowired ObjectMapper objectMapper;
@@ -51,21 +55,18 @@ class ExerciseCommandIntegrationTest extends IntegrationTestBase {
     @Autowired GuestRepository guestRepository;
 
     private Member manager;
-    private Member subManager;
     private Member normalMember;
     private Party party;
 
     @BeforeEach
     void setUp() {
         manager = memberRepository.save(MemberFixture.createMember("모임장", Gender.MALE, Level.A, 1001L, LocalDate.of(2000, 1, 1)));
-        subManager = memberRepository.save(MemberFixture.createMember("부모임장", Gender.FEMALE, Level.B, 1002L, LocalDate.of(2000, 1, 1)));
         normalMember = memberRepository.save(MemberFixture.createMember("일반멤버", Gender.MALE, Level.C, 1003L, LocalDate.of(2000, 1, 1)));
 
         PartyAddr addr = partyAddrRepository.save(PartyFixture.createPartyAddr("서울특별시", "강남구"));
         party = partyRepository.save(PartyFixture.createParty("테스트 모임", manager.getId(), addr));
 
         memberPartyRepository.save(MemberFixture.createMemberParty(party, manager, Role.PARTY_MANAGER));
-        memberPartyRepository.save(MemberFixture.createMemberParty(party, subManager, Role.PARTY_SUBMANAGER));
         memberPartyRepository.save(MemberFixture.createMemberParty(party, normalMember, Role.PARTY_MEMBER));
     }
 
@@ -319,6 +320,86 @@ class ExerciseCommandIntegrationTest extends IntegrationTestBase {
                         .andExpect(status().isForbidden())
                         .andExpect(jsonPath("$.code").value(ExerciseErrorCode.GUEST_NOT_INVITED_BY_MEMBER.getCode()))
                         .andExpect(jsonPath("$.message").value(ExerciseErrorCode.GUEST_NOT_INVITED_BY_MEMBER.getMessage()));
+            }
+        }
+    }
+
+    @Nested
+    @DisplayName("GET /api/exercises/{exerciseId}/guests - 내가 초대한 운동 게스트 조회")
+    class GetMyInvitedGuests {
+
+        private Exercise exercise;
+
+        @BeforeEach
+        void setUp() {
+            exercise = exerciseRepository.save(
+                    ExerciseFixture.createExerciseWithAddr(party, LocalDate.now().minusDays(1), 1));
+        }
+
+        @Nested
+        @DisplayName("성공 케이스")
+        class Success {
+
+            @Test
+            @DisplayName("내가_초대한_게스트만_참가번호와_대기상태와_함께_반환된다")
+            void 내가_초대한_게스트만_참가번호와_대기상태와_함께_반환된다() throws Exception {
+                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
+
+                var myFirstGuest = GuestFixture.createGuest(exercise, manager.getId(), "내게스트1", Gender.MALE);
+                guestRepository.save(myFirstGuest);
+
+                var otherGuest = GuestFixture.createGuest(exercise, normalMember.getId(), "다른사람게스트", Gender.MALE);
+                guestRepository.save(otherGuest);
+
+                var mySecondGuest = GuestFixture.createGuest(exercise, manager.getId(), "내게스트2", Gender.FEMALE);
+                guestRepository.save(mySecondGuest);
+
+                mockMvc.perform(get("/api/exercises/{exerciseId}/guests", exercise.getId()))
+                        .andExpect(status().isOk())
+                        .andExpect(jsonPath("$.data.totalCount").value(2))
+                        .andExpect(jsonPath("$.data.maleCount").value(1))
+                        .andExpect(jsonPath("$.data.femaleCount").value(1))
+                        .andExpect(jsonPath("$.data.list[0].guestId").isNumber())
+                        .andExpect(jsonPath("$.data.list[0].isWaiting").value(false))
+                        .andExpect(jsonPath("$.data.list[0].participantNumber").value(1))
+                        .andExpect(jsonPath("$.data.list[0].name").value("내게스트1"))
+                        .andExpect(jsonPath("$.data.list[0].gender").value("MALE"))
+                        .andExpect(jsonPath("$.data.list[0].level").value("B"))
+                        .andExpect(jsonPath("$.data.list[0].inviterName").value(manager.getMemberName()))
+                        .andExpect(jsonPath("$.data.list[1].guestId").isNumber())
+                        .andExpect(jsonPath("$.data.list[1].isWaiting").value(true))
+                        .andExpect(jsonPath("$.data.list[1].participantNumber").value(2))
+                        .andExpect(jsonPath("$.data.list[1].name").value("내게스트2"))
+                        .andExpect(jsonPath("$.data.list[1].gender").value("FEMALE"))
+                        .andExpect(jsonPath("$.data.list[1].level").value("B"))
+                        .andExpect(jsonPath("$.data.list[1].inviterName").value(manager.getMemberName()));
+            }
+        }
+
+        @Nested
+        @DisplayName("실패 케이스")
+        class Failure {
+
+            @Test
+            @DisplayName("존재하지_않는_운동이면_에러를_반환한다")
+            void 존재하지_않는_운동이면_에러를_반환한다() throws Exception {
+                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
+
+                mockMvc.perform(get("/api/exercises/{exerciseId}/guests", 999L))
+                        .andExpect(status().isNotFound())
+                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getCode()))
+                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getMessage()));
+            }
+
+            @Test
+            @DisplayName("존재하지_않는_멤버면_에러를_반환한다")
+            void 존재하지_않는_멤버면_에러를_반환한다() throws Exception {
+                SecurityContextHelper.setAuthentication(999L, "없는멤버");
+
+                mockMvc.perform(get("/api/exercises/{exerciseId}/guests", exercise.getId()))
+                        .andExpect(status().isNotFound())
+                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getCode()))
+                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getMessage()));
             }
         }
     }
