+++ src/test/java/umc/cockple/demo/domain/exercise/integration/ExerciseQueryIntegrationTest.java
@@ -93,86 +93,6 @@ class ExerciseQueryIntegrationTest extends IntegrationTestBase {
     }
 
     @Nested
-    @DisplayName("GET /api/exercises/{exerciseId}/guests - 내가 초대한 운동 게스트 조회")
-    class GetMyInvitedGuests {
-
-        private Exercise exercise;
-
-        @BeforeEach
-        void setUp() {
-            exercise = exerciseRepository.save(
-                    ExerciseFixture.createExerciseWithAddr(party, LocalDate.now().minusDays(1), 1));
-        }
-
-        @Nested
-        @DisplayName("성공 케이스")
-        class Success {
-
-            @Test
-            @DisplayName("내가_초대한_게스트만_참가번호와_대기상태와_함께_반환된다")
-            void 내가_초대한_게스트만_참가번호와_대기상태와_함께_반환된다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                var myFirstGuest = GuestFixture.createGuest(exercise, manager.getId(), "내게스트1", Gender.MALE);
-                guestRepository.save(myFirstGuest);
-
-                var otherGuest = GuestFixture.createGuest(exercise, normalMember.getId(), "다른사람게스트", Gender.MALE);
-                guestRepository.save(otherGuest);
-
-                var mySecondGuest = GuestFixture.createGuest(exercise, manager.getId(), "내게스트2", Gender.FEMALE);
-                guestRepository.save(mySecondGuest);
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}/guests", exercise.getId()))
-                        .andExpect(status().isOk())
-                        .andExpect(jsonPath("$.data.totalCount").value(2))
-                        .andExpect(jsonPath("$.data.maleCount").value(1))
-                        .andExpect(jsonPath("$.data.femaleCount").value(1))
-                        .andExpect(jsonPath("$.data.list[0].guestId").isNumber())
-                        .andExpect(jsonPath("$.data.list[0].isWaiting").value(false))
-                        .andExpect(jsonPath("$.data.list[0].participantNumber").value(1))
-                        .andExpect(jsonPath("$.data.list[0].name").value("내게스트1"))
-                        .andExpect(jsonPath("$.data.list[0].gender").value("MALE"))
-                        .andExpect(jsonPath("$.data.list[0].level").value("B"))
-                        .andExpect(jsonPath("$.data.list[0].inviterName").value(manager.getMemberName()))
-                        .andExpect(jsonPath("$.data.list[1].guestId").isNumber())
-                        .andExpect(jsonPath("$.data.list[1].isWaiting").value(true))
-                        .andExpect(jsonPath("$.data.list[1].participantNumber").value(2))
-                        .andExpect(jsonPath("$.data.list[1].name").value("내게스트2"))
-                        .andExpect(jsonPath("$.data.list[1].gender").value("FEMALE"))
-                        .andExpect(jsonPath("$.data.list[1].level").value("B"))
-                        .andExpect(jsonPath("$.data.list[1].inviterName").value(manager.getMemberName()));
-            }
-        }
-
-        @Nested
-        @DisplayName("실패 케이스")
-        class Failure {
-
-            @Test
-            @DisplayName("존재하지_않는_운동이면_에러를_반환한다")
-            void 존재하지_않는_운동이면_에러를_반환한다() throws Exception {
-                SecurityContextHelper.setAuthentication(manager.getId(), manager.getNickname());
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}/guests", 999L))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.EXERCISE_NOT_FOUND.getMessage()));
-            }
-
-            @Test
-            @DisplayName("존재하지_않는_멤버면_에러를_반환한다")
-            void 존재하지_않는_멤버면_에러를_반환한다() throws Exception {
-                SecurityContextHelper.setAuthentication(999L, "없는멤버");
-
-                mockMvc.perform(get("/api/exercises/{exerciseId}/guests", exercise.getId()))
-                        .andExpect(status().isNotFound())
-                        .andExpect(jsonPath("$.code").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getCode()))
-                        .andExpect(jsonPath("$.message").value(ExerciseErrorCode.MEMBER_NOT_FOUND.getMessage()));
-            }
-        }
-    }
-
-    @Nested
     @DisplayName("GET /api/parties/{partyId}/exercises/calender - 모임 운동 캘린더 조회")
     class GetPartyExerciseCalendar {
